internal: log inline-rescue squashed exceptions at level TRACE
Replaces all instances of `rescue nil` (which silently squashes any exception
that is raised in the preceeding statement) with `resque log_and_squash`, which
will log the exception at level TRACE before emitting `nil`.

The underlying behaviour is unchanged, but when TRACE-level logging is enabled,
it becomes easier to see what is going on in the application.

Fixes #43